# openapi.yaml
openapi: 3.0.3
info:
  title: Last Price
  description: |
    Server for conducting A/B testing experiments on pricing strategies using Last Price integration.
    
    This server provides endpoints for:
    - Assigning users to experiment variants
    - Tracking pricing page views and conversions
    - Processing Paid.ai webhook events
    - Analyzing experiment results
    
    **Architecture**: The system uses deterministic SHA-256 hashing to ensure consistent user-to-variant mapping,
    emits signals to Paid.ai's Signals API for event tracking, and provides real-time analytics.
  version: 1.0.0
  contact:
    name: API Support
  license:
    name: MIT
servers:
  - url: http://localhost:3000
    description: Local development server
  - url: https://api.example.com
    description: Production server
...
paths:
...
  /customers/{customerId}:
    get:
      tags:
        - Customers
      summary: Retrieve a customer
      description: Retrieves a customer by ID from Last Price
      operationId: getLastPriceCustomer
      parameters:
        - name: customerId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LastPriceCustomer'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    put:
      tags:
        - Customers
      summary: Update a customer
      description: Updates a customer in Last Price
      operationId: updateLastPriceCustomer
      parameters:
        - name: customerId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LastPriceCustomerInput'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LastPriceCustomer'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      tags:
        - Customers
      summary: Delete a customer
      description: Deletes a customer from Last Price
      operationId: deleteLastPriceCustomer
      parameters:
        - name: customerId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Deleted
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /customers/external/{externalId}:
    get:
      tags:
        - Customers
      summary: Retrieve customer by external ID
      description: Retrieves a customer by external ID from Last Price
      operationId: getLastPriceCustomerByExternalId
      parameters:
        - name: externalId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LastPriceCustomer'

    put:
      tags:
        - Customers
      summary: Update customer by external ID
      description: Updates a customer by external ID in Last Price
      operationId: updateLastPriceCustomerByExternalId
      parameters:
        - name: externalId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LastPriceCustomerInput'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LastPriceCustomer'

    delete:
      tags:
        - Customers
      summary: Delete customer by external ID
      description: Deletes a customer by external ID from Last Price
      operationId: deleteLastPriceCustomerByExternalId
      parameters:
        - name: externalId
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Deleted
...
  /credits:
    get:
      tags:
        - Credits
      summary: List all credit bundles
      description: Retrieves all credit bundles from Last Price
      operationId: listLastPriceCredits
      parameters:
        - name: customerId
          in: query
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Retrieved
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/LastPriceCreditBundle'

    post:
      tags:
        - Credits
      summary: Create a credit bundle
      description: Creates a new credit bundle in Last Price
      operationId: createLastPriceCredit
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LastPriceCreditBundleInput'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LastPriceCreditBundle'
  /credits/{creditId}:
    get:
      tags:
        - Credits
      summary: Retrieve a credit bundle
      description: Retrieves a credit bundle by ID from Last Price
      operationId: getLastPriceCredit
      parameters:
        - name: creditId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LastPriceCreditBundle'

    delete:
      tags:
        - Credits
      summary: Delete a credit bundle
      description: Deletes a credit bundle from Last Price
      operationId: deleteLastPriceCredit
      parameters:
        - name: creditId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Deleted
  /credits/balance:
    get:
      tags:
        - Credits
      summary: Get customer credit balance
      description: Retrieves the current credit balance for a customer
      operationId: getLastPriceCreditBalance
      parameters:
        - name: customerId
          in: query
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  customerId:
                    type: string
                    format: uuid
                  balance:
                    type: number
                  currency:
                    type: string
  /costs:
    get:
      tags:
        - Costs
      summary: List all costs
      description: Retrieves cost tracking data from Last Price
      operationId: listLastPriceCosts
      parameters:
        - name: startDate
          in: query
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          schema:
            type: string
            format: date
        - name: agentId
          in: query
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Retrieved
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/LastPriceCost'
...
components:
  schemas:
...
    LastPriceAddress:
      type: object
      properties:
        line1:
          type: string
          example: "123 Business Ave"
        line2:
          type: string
          example: "Suite 100"
        city:
          type: string
          example: "San Francisco"
        state:
          type: string
          example: "CA"
        zipCode:
          type: string
          example: "94105"
        country:
          type: string
          example: "USA"
    LastPriceCustomer:
      type: object
      properties:
        id:
          type: string
          format: uuid
        organizationId:
          type: string
          format: uuid
        name:
          type: string
          example: "Acme Corporation"
        phone:
          type: string
          example: "+1-555-0123"
        website:
          type: string
          example: "https://acme.com"
        employeeCount:
          type: integer
          example: 100
        annualRevenue:
          type: number
          example: 1000000
        taxExemptStatus:
          type: string
          enum: [none, exempt, reverse]
          default: none
        creationSource:
          type: string
          enum: [manual, api, crm, other]
        externalId:
          type: string
          example: "acme_123"
        creationState:
          type: string
          example: "active"
        billingAddress:
          $ref: '#/components/schemas/LastPriceAddress'
    LastPriceCustomerInput:
      type: object
      required:
        - name
      properties:
        name:
          type: string
        phone:
          type: string
        website:
          type: string
        employeeCount:
          type: integer
        annualRevenue:
          type: number
        taxExemptStatus:
          type: string
          enum: [none, exempt, reverse]
        creationSource:
          type: string
          enum: [manual, api, crm, other]
        externalId:
          type: string
        billingAddress:
          $ref: '#/components/schemas/LastPriceAddress'
...
    LastPriceCreditBundle:
      type: object
      properties:
        id:
          type: string
          format: uuid
        organizationId:
          type: string
          format: uuid
        customerId:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        amount:
          type: number
        currency:
          type: string
        remainingBalance:
          type: number
        expiresAt:
          type: string
          format: date-time
        status:
          type: string
          enum: [active, expired, exhausted]
        createdAt:
          type: string
          format: date-time
    LastPriceCreditBundleInput:
      type: object
      required:
        - customerId
        - amount
      properties:
        customerId:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        amount:
          type: number
        currency:
          type: string
          default: USD
        expiresAt:
          type: string
          format: date-time

    # ============ Cost Schemas ============
    LastPriceCost:
      type: object
      properties:
        id:
          type: string
          format: uuid
        organizationId:
          type: string
          format: uuid
        agentId:
          type: string
          format: uuid
        customerId:
          type: string
          format: uuid
        costType:
          type: string
          enum: [compute, storage, api_call, data_transfer, other]
        amount:
          type: number
        currency:
          type: string
        units:
          type: number
        unitPrice:
          type: number
        metadata:
          type: object
          additionalProperties: true
        timestamp:
          type: string
          format: date-time
...
    Error:
      type: object
      properties:
        error:
          type: string
          example: Invalid request parameters
        message:
          type: string
          example: userId must be a non-empty string
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Bearer token for Last Price API authentication
    webhookSignature:
      type: apiKey
      in: header
      name: x-paid-signature
      description: HMAC signature for webhook verification